fixing build task and comments

+++ obc/app/drivers/arducam/arducam.h
@@ -230,4 +230,4 @@ obc_error_code_t camReadSensorReg16_8(uint32_t regID, uint8_t* regDat);
  * @param reglist List of registers and data to write
  * @return Error code indicating if the writes were successful
  */
-obc_error_code_t camWriteSensorRegs16_8(const sensor_config_t reglist[], uint16_t reglistLen);
+obc_error_code_t camWriteSensorRegs16_8(const sensor_config_t reglist[], size_t reglistLen);
